Make student component a container.
